backend/requirements.txt
# FastAPI & Server
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.4
hiredis==2.3.2

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4

# Azure SDK
azure-identity==1.16.0
azure-keyvault-secrets==4.8.0
azure-eventhub==5.11.6
azure-cosmos==4.5.1
azure-ai-ml==1.15.0
azure-monitor-opentelemetry==1.3.0

# OpenTelemetry
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-instrumentation-redis==0.45b0

# ML
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.1
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2

# HTTP Client
httpx==0.27.0
aiohttp==3.9.5

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.3.0
